time fixes

+++ src/main/java/com/dugonggames/shooter/shooter/Ship.java
@@ -15,7 +15,7 @@ import lombok.Value;
 public class Ship {
     MovingPoint location;
     int health;
-    int missileTimer;
+    double missileTimer;
 
     public void draw(GfxWrapper gfx, double angle){
         double x = location.location.x;
